pagination/filters.py
from .models import Data
import django_filters; 

class DataFilter(django_filters.FilterSet):
    name = django_filters.CharFilter(field_name='name',lookup_expr='icontains')
    
    class Meta:
        model=Data
        fields=['name']
        
    