requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
crispy-bootstrap4==2025.6
Django==6.0
django-crispy-forms==2.5
django-dotenv==1.4.2
django-filter==25.2
idna==3.11
pillow==12.0.0
requests==2.32.5
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.2
